+++ packages/core/src/hook/index.ts
@@ -1,6 +1,5 @@
 export * from "./hierarchy";
 export * from "./hook";
-export * from "./hook-call";
 export * from "./hook-collector";
 export * from "./hook-name";
 export * from "./is";