+++ packages/plugins/eslint-plugin-react-hooks/src/rules/ensure-use-callback-has-non-empty-deps.ts
@@ -1,5 +1,5 @@
 import { findVariableByNameUpToGlobal, getVariableInit, is, NodeType } from "@eslint-react/ast";
-import { isUseCallbackCall, unsafeIsReactHookCall } from "@eslint-react/core";
+import { isReactHookCall, isUseCallbackCall } from "@eslint-react/core";
 import { getPragmaFromContext } from "@eslint-react/jsx";
 import { F, O } from "@eslint-react/tools";
 import type { ESLintUtils } from "@typescript-eslint/utils";
@@ -33,7 +33,7 @@ export default createRule<[], MessageID>({
       CallExpression(node) {
         const initialScope = context.sourceCode.getScope?.(node) ?? context.getScope();
 
-        if (!unsafeIsReactHookCall(node) || !isUseCallbackCall(node, context, pragma)) {
+        if (!isReactHookCall(node) || !isUseCallbackCall(node, context, pragma)) {
           return;
         }