+++ packages/plugins/eslint-plugin-react-hooks/src/rules/prefer-use-state-lazy-initialization.ts
@@ -1,6 +1,6 @@
 // Ported from https://github.com/jsx-eslint/eslint-plugin-react/pull/3579/commits/ebb739a0fe99a2ee77055870bfda9f67a2691374
 import { getNestedCallExpressions } from "@eslint-react/ast";
-import { isUseStateCall, unsafeIsReactHookCall } from "@eslint-react/core";
+import { isReactHookCall, isUseStateCall } from "@eslint-react/core";
 import { getPragmaFromContext } from "@eslint-react/jsx";
 import type { ESLintUtils } from "@typescript-eslint/utils";
 import { type ConstantCase } from "string-ts";
@@ -35,7 +35,7 @@ export default createRule<[], MessageID>({
 
     return {
       CallExpression(node) {
-        if (!unsafeIsReactHookCall(node) || !isUseStateCall(node, context, pragma)) {
+        if (!isReactHookCall(node) || !isUseStateCall(node, context, pragma)) {
           return;
         }